Added unRadio v2
New RADIO.DAT tool. Parse better the data and extracts added chars as tga files.
* Added the @jeromegrosse character maps. (Big thanks)
IMPORTANT: those chars must be parsed in some way, because isn't the same in all version. This mean the bytes "9601" not always is "気" but I'm working on this. I'll add this in a few days.